requirements.txt
aenum==2.2.3
appdirs==1.4.4
attrs==19.3.0
bcrypt==3.2.0
beautifulsoup4==4.9.1
blinker==1.4
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
distlib==0.3.1
dnspython==2.1.0
email-validator==1.1.2
facebook-sdk==3.1.0
Faker==4.1.1
fbchat==1.9.7
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-json-multidict==1.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
numpy==1.19.0
paho-mqtt==1.5.0
pandas==1.0.5
phonenumbers==8.12.15
Pillow==8.1.0
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
PuLP==2.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.22
style==1.1.0
text-unidecode==1.3
tika==1.24
update==0.0.1
urllib3==1.25.9
virtualenv==20.2.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
WTForms==2.3.3
